+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Versionning adheres to [Semantic Versioning] (https://semver.org/spec/v2.0.0.htm
   I change the previous behaviour as the new metadata fetching feature need
   probably some more love before being blindly trusted.
 - Rename `get` command to `info`and `update` command to `edit`.
+- Merge `list` and `search` commands.
 - Replace the Makefile, which was over-enginneered (and probably badly done),
   by a small shell script that wraps go binary to supply version information
   and manpage generation/installation